ci_requirements.txt
pyyaml>=6.0
requests>=2.28.0
pysigma>=0.10.0
pysigma-backend-elasticsearch>=1.0.0
pysigma-pipeline-sysmon>=1.0.0
jsonschema>=4.17.0
elasticsearch>=8.0.0
colorama>=0.4.6